Move /bug-bounty under /get-involved/
